+++ mzdb-access/src/main/java/fr/profi/mzdb/io/reader/bb/BoundingBoxBuilder.java
@@ -24,10 +24,10 @@ public class BoundingBoxBuilder {
 	public static BoundingBox buildBB(
 		int bbId,
 		byte[] bytes,
-		int firstScanId,
-		int lastScanId,
-		Map<Integer, ScanHeader> scanHeaderById,
-		Map<Integer, DataEncoding> dataEncodingByScanId
+		long firstScanId,
+		long lastScanId,
+		Map<Long, ScanHeader> scanHeaderById,
+		Map<Long, DataEncoding> dataEncodingByScanId
 	) throws StreamCorruptedException {
 		return new BoundingBox(bbId, new BytesReader(bytes, firstScanId, lastScanId, scanHeaderById, dataEncodingByScanId));
 	}
@@ -35,10 +35,10 @@ public class BoundingBoxBuilder {
 	public static BoundingBox buildBB(
 		int bbId,
 		SQLiteBlob blob,
-		int firstScanId,
-		int lastScanId,
-		Map<Integer, ScanHeader> scanHeaderById,
-		Map<Integer, DataEncoding> dataEncodingByScanId
+		long firstScanId,
+		long lastScanId,
+		Map<Long, ScanHeader> scanHeaderById,
+		Map<Long, DataEncoding> dataEncodingByScanId
 	) throws StreamCorruptedException {
 		return new BoundingBox(bbId, new SQLiteBlobReader(blob, firstScanId, lastScanId, scanHeaderById, dataEncodingByScanId) );
 	}
@@ -46,10 +46,10 @@ public class BoundingBoxBuilder {
 	public static BoundingBox buildBB(
 		int bbId,
 		InputStream stream,
-		int firstScanId,
-		int lastScanId,
-		Map<Integer, ScanHeader> scanHeaderById,
-		Map<Integer, DataEncoding> dataEncodingByScanId
+		long firstScanId,
+		long lastScanId,
+		Map<Long, ScanHeader> scanHeaderById,
+		Map<Long, DataEncoding> dataEncodingByScanId
 	) {
 		return new BoundingBox(bbId, new StreamReader(stream, firstScanId, lastScanId, scanHeaderById, dataEncodingByScanId) );
 	}
