+++ mzdb-access/src/main/java/fr/profi/mzdb/io/reader/bb/BytesReader.java
@@ -36,10 +36,10 @@ public class BytesReader extends AbstractBlobReader {
 	 */
 	public BytesReader(
 		byte[] bytes,
-		int firstScanId,
-		int lastScanId,
-		Map<Integer, ScanHeader> scanHeaderById,
-		Map<Integer, DataEncoding> dataEncodingByScanId
+		long firstScanId,
+		long lastScanId,
+		Map<Long, ScanHeader> scanHeaderById,
+		Map<Long, DataEncoding> dataEncodingByScanId
 	) throws StreamCorruptedException {
 		super(firstScanId, lastScanId, scanHeaderById, dataEncodingByScanId);
