+++ mzdb-access/src/main/java/fr/profi/mzdb/io/reader/bb/StreamReader.java
@@ -33,10 +33,10 @@ public class StreamReader extends AbstractBlobReader {
 	 */
 	public StreamReader(
 		InputStream inputStream,
-		int firstScanId,
-		int lastScanId,
-		Map<Integer, ScanHeader> scanHeaderById,
-		Map<Integer, DataEncoding> dataEncodingByScanId
+		long firstScanId,
+		long lastScanId,
+		Map<Long, ScanHeader> scanHeaderById,
+		Map<Long, DataEncoding> dataEncodingByScanId
 	) {
 		super(firstScanId, lastScanId, scanHeaderById, dataEncodingByScanId);
